cleaning project and keeping the base
