+++ Assets/Prefabs/Player.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 7246502344431888075}
   - component: {fileID: 8549289276102554192}
   - component: {fileID: 956186952977780209}
-  - component: {fileID: 428389240537827745}
   - component: {fileID: 2015712715758816506}
   - component: {fileID: 4220220450115104865}
   - component: {fileID: 601201847686730596}
@@ -131,24 +130,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   ShowTopMostFoldoutHeaderGroup: 1
-  player: {fileID: 428389240537827745}
   spawnedObjectPrefab: {fileID: 1163429242141364165, guid: 5fcc5d08877d4124491aded62a75fe3d, type: 3}
   moveSpeed: 3
---- !u!114 &428389240537827745
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3258933237172009959}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: adc0028cbdb1d6548869f915271cc67b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ShowTopMostFoldoutHeaderGroup: 1
-  HP: 1000
-  pseudo: Ixenooz
 --- !u!114 &2015712715758816506
 MonoBehaviour:
   m_ObjectHideFlags: 0
