--- Assets/Scripts/Player.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: adc0028cbdb1d6548869f915271cc67b
